+++ public/js/script.js
@@ -1,9 +1,9 @@
-import { EventListeners } from './event.js'
+import { EventListener } from './event.js'
 
 function qS(selector) {
   return document.querySelector(selector);
 }
 
-qS('header input').addEventListener('keydown', () => {EventListeners.search(event)})
+qS('header input').addEventListener('keydown', () => {EventListener.search(event)})
 
-qS('main').addEventListener('click', ()  => {EventListeners.forks(event)})
+qS('main').addEventListener('click', ()  => {EventListener.forks(event)})
